Refactor react-query to reduce code repetition
