+++ projects/packages/my-jetpack/_inc/data/constants.ts
@@ -1,4 +1,3 @@
 const REST_API_NAMESPACE = 'my-jetpack/v1';
 
 export const REST_API_SITE_PURCHASES_ENDPOINT = `${ REST_API_NAMESPACE }/site/purchases`;
-export const REST_API_SITE_PURCHASES_QUERY_KEY = 'purchases';
